requirements.txt
Pillow==11.3.0
openpyxl==3.1.5
tencentcloud-sdk-python-common
tencentcloud-sdk-python-ocr

xlwings==0.33.15
python-dotenv==1.1.1

pywin32==311; sys_platform == "win32"
